backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
langchain==0.0.335
langchain-community==0.0.13
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==1.3.5
tiktoken==0.5.1
